requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.7
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
CodeConvert==3.0.2
colorama==0.4.5
coreapi==2.3.3
coreschema==0.0.4
Deprecated==1.2.13
Django==4.0.5
django-appconf==1.0.5
django-celery-email==3.0.0
django-models-ext==1.1.10
django-redis==5.2.0
django-six==1.0.4
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.0
dnspython==2.2.1
drf-yasg==1.21.3
eventlet==0.33.1
furl==2.1.3
greenlet==1.1.2
idna==3.3
inflection==0.5.1
isoweek==1.3.3
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.1
orderedmultidict==1.0.1
packaging==21.3
Pillow==9.1.1
prompt-toolkit==3.0.30
psycopg2==2.9.3
PyJWT==1.7.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-decouple==3.6
sqlparse==0.4.2
TimeConvert==2.0.3
tzdata==2022.1
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.10
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
xlwt==1.3.0
